% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTn5Bias.R
\name{getTn5Bias}
\alias{getTn5Bias}
\title{getTn5Bias}
\usage{
getTn5Bias(convert = TRUE, includeNs = TRUE)
}
\arguments{
\item{convert}{Logical; whether to convert into an (observed/expected) odds 
ratio (default TRUE). If FALSE, the bias values as stored, i.e. rounded, 
100*log2(OR) to ease the memory footprint.}

\item{includeNs}{Logical; whether to include kmers with leading or trailing 
Ns (default TRUE).}
}
\value{
A data.frame with the columns kmer and bias
}
\description{
Returns a table of Tn5 bias per 10-mers, obtained from SELMA (Hu et al. 2022)
 using ATAC-seq on naked human DNA.
}
\examples{
bias <- getTn5Bias()
head(bias)
}
\references{
Hu, S.S., Liu, L., Li, Q. et al. Intrinsic bias estimation for improved 
analysis of bulk and single-cell chromatin accessibility profiles using 
SELMA. Nat Commun 13, 5533 (2022). https://doi.org/10.1038/s41467-022-33194-z
}
