% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tn5bias}
\alias{tn5bias}
\title{Tn5 bias}
\format{
An object of class \code{integer} of length 1746944.
}
\usage{
tn5bias
}
\description{
Raw Tn5 bias per 10-mers obtained using SELMA (Hu et al. 2022),
downloaded from the SELMA repository. For low memory footprint
this has been saved as integers without the kmers, using
system.file('inst/scripts/', 'bias_data_gen.R', package='tn5bias')
Use \link{getTn5Bias} to obtain an interpretable kmer-bias table.
}
\keyword{datasets}
