% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tn5biasMotif}
\alias{tn5biasMotif}
\title{Tn5 bias motif}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 4 rows and 10 columns.
}
\usage{
tn5biasMotif
}
\description{
A motif summary of the Tn5 bias, summarized from the kmer-level bias
obtained using SELMA (Hu et al. 2022). See \link{getTn5Bias} for more
information. This has been generated using
system.file('inst/scripts/', 'biasMotif.R', package='tn5bias')
}
\keyword{datasets}
