% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMotifTn5Bias.R
\name{getMotifTn5Bias}
\alias{getMotifTn5Bias}
\title{getMotifTn5Bias}
\usage{
getMotifTn5Bias(x, bias = NULL, ...)
}
\arguments{
\item{x}{A DNA motif, in any format recognized by 
\link[universalmotif]{convert_motifs}.}

\item{bias}{An optional 10-mer bias table, with the columns 'kmer' and 
'bias'. By default, an (internal) table derived from SELMA will be used 
(see \link{getTn5Bias}).}

\item{...}{Any argument passed to \code{\link{getSequenceTn5Bias}}.}
}
\value{
A vector of bias per position.
}
\description{
Returns the per-position Tn5 bias of a DNA motif.
}
\examples{
# get an example motif
jaspar <- universalmotif::read_jaspar(system.file("extdata", "jaspar.txt",
                                                  package="universalmotif"))
getMotifTn5Bias(jaspar[[1]])
}
