% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSequenceTn5Bias.R
\name{getSequenceTn5Bias}
\alias{getSequenceTn5Bias}
\title{getSequenceTn5Bias}
\usage{
getSequenceTn5Bias(
  x,
  bias = NULL,
  nthreads = 1L,
  complement = FALSE,
  asInteger = FALSE,
  shift = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{x}{A (set of) DNA sequence(s), either as a \link[Rsamtools]{FaFile} 
object, a character vector, or a DNAString or DNAStringSet.}

\item{bias}{An optional 10-mer bias table, with the columns 'kmer' and 
'bias'. By default, an (internal) table derived from SELMA will be used 
(see \link{getTn5Bias}).}

\item{nthreads}{The number of threads to use (default 1). Alternatively,}

\item{complement}{Logical; whether to compute bias for the complement of `x`.
This is useful to get the bias for the negative strand.}

\item{asInteger}{Logical; whether to convert the (observed/expected) odds 
ratio to rounded 100*log2(OR) to ease the memory footprint. Recommended when
computing genome-wide bias.}

\item{shift}{Logical; whether to shift the bias, so that it represents not the 5'
ends of (positives) fragments, but the insertion site (midpoint between the
cleavages sites on each strand). The shifted bias is consistent with the 
shifting often performed on ATAC-seq data, and is thus enabled by default.
Note, however, that this is different from the original SELMA methods.}

\item{verbose}{Logical; whether to output progress messages.}
}
\value{
An AtomicList of bias per position.
}
\description{
getSequenceTn5Bias
}
\examples{
fakeSequence <- "NNNNNACGTACGTGCGGGCTATGTCACGTNNNNNNN"
getSequenceTn5Bias(fakeSequence)
}
